+++ magics.txt
