+++ src/Search/search.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../Board/board.h"
+#include "../Core/board.h"
 
 double minimax(Board *b, int depth, bool maximiser);
 Move search(Board b, int depth);//search uses minimax but returns a move instead of an evaluation
