+++ src/main.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
-#include "Board/board.h"
+#include "Core/board.h"
 #include "ui/ui.h"
-#include "Search/movegen.h"
-#include "Search/Magic/magic.h"
+#include "Movegen/movegen.h"
+#include "Movegen/Magic/magic.h"
 
 int main() {
   printf("[creating board...]\n");
