+++ src/ui/print.c
@@ -1,7 +1,7 @@
 #include "print.h"
 #include <stdio.h>
 #include <string.h>
-#include "../Board/Bitboards/bitboard.h"
+#include "../Core/bitboard.h"
 
 void setLightColor(printSettings *settings, const char *colorID){
   settings->lightColor[0] = 0;
